+++ keymap.c
@@ -1149,7 +1149,7 @@ void km_error_key(enum MenuType menu)
  * mutt_parse_push - Parse the 'push' command - Implements Command::parse()
  */
 enum CommandResult mutt_parse_push(struct Buffer *buf, struct Buffer *s,
-                                   unsigned long data, struct Buffer *err)
+                                   intptr_t data, struct Buffer *err)
 {
   mutt_extract_token(buf, s, MUTT_TOKEN_CONDENSE);
   if (MoreArgs(s))
@@ -1312,7 +1312,7 @@ const struct Binding *km_get_table(enum MenuType menu)
  * bind menu-name `<key_sequence>` function-name
  */
 enum CommandResult mutt_parse_bind(struct Buffer *buf, struct Buffer *s,
-                                   unsigned long data, struct Buffer *err)
+                                   intptr_t data, struct Buffer *err)
 {
   const struct Binding *bindings = NULL;
   enum MenuType menu[sizeof(Menus) / sizeof(struct Mapping) - 1];
@@ -1453,7 +1453,7 @@ static void km_unbind_all(struct Keymap **map, unsigned long mode)
  * unbind `<menu-name[,...]|*>` [`<key_sequence>`]
  */
 enum CommandResult mutt_parse_unbind(struct Buffer *buf, struct Buffer *s,
-                                     unsigned long data, struct Buffer *err)
+                                     intptr_t data, struct Buffer *err)
 {
   bool menu[MENU_MAX] = { 0 };
   bool all_keys = false;
@@ -1518,7 +1518,7 @@ enum CommandResult mutt_parse_unbind(struct Buffer *buf, struct Buffer *s,
  * macro `<menu>` `<key>` `<macro>` `<description>`
  */
 enum CommandResult mutt_parse_macro(struct Buffer *buf, struct Buffer *s,
-                                    unsigned long data, struct Buffer *err)
+                                    intptr_t data, struct Buffer *err)
 {
   enum MenuType menu[sizeof(Menus) / sizeof(struct Mapping) - 1];
   int num_menus = 0;
@@ -1572,7 +1572,7 @@ enum CommandResult mutt_parse_macro(struct Buffer *buf, struct Buffer *s,
  * mutt_parse_exec - Parse the 'exec' command - Implements Command::parse()
  */
 enum CommandResult mutt_parse_exec(struct Buffer *buf, struct Buffer *s,
-                                   unsigned long data, struct Buffer *err)
+                                   intptr_t data, struct Buffer *err)
 {
   int ops[128];
   int nops = 0;
