+++ score.c
@@ -100,7 +100,7 @@ void mutt_check_rescore(struct Mailbox *m)
  * mutt_parse_score - Parse the 'score' command - Implements Command::parse()
  */
 enum CommandResult mutt_parse_score(struct Buffer *buf, struct Buffer *s,
-                                    unsigned long data, struct Buffer *err)
+                                    intptr_t data, struct Buffer *err)
 {
   struct Score *ptr = NULL, *last = NULL;
   char *pattern = NULL, *pc = NULL;
@@ -204,7 +204,7 @@ void mutt_score_message(struct Mailbox *m, struct Email *e, bool upd_mbox)
  * mutt_parse_unscore - Parse the 'unscore' command - Implements Command::parse()
  */
 enum CommandResult mutt_parse_unscore(struct Buffer *buf, struct Buffer *s,
-                                      unsigned long data, struct Buffer *err)
+                                      intptr_t data, struct Buffer *err)
 {
   struct Score *tmp = NULL, *last = NULL;
