+++ score.h
@@ -36,8 +36,8 @@ extern short C_ScoreThresholdFlag;
 extern short C_ScoreThresholdRead;
 
 void mutt_check_rescore(struct Mailbox *m);
-enum CommandResult mutt_parse_score(struct Buffer *buf, struct Buffer *s, unsigned long data, struct Buffer *err);
-enum CommandResult mutt_parse_unscore(struct Buffer *buf, struct Buffer *s, unsigned long data, struct Buffer *err);
+enum CommandResult mutt_parse_score(struct Buffer *buf, struct Buffer *s, intptr_t data, struct Buffer *err);
+enum CommandResult mutt_parse_unscore(struct Buffer *buf, struct Buffer *s, intptr_t data, struct Buffer *err);
 void mutt_score_message(struct Mailbox *m, struct Email *e, bool upd_ctx);
 
 #endif /* MUTT_SCORE_H */
